Add Green and Red Balls
The FlyingFish has to catch colored Balls to increase the game score. Now that we had coded the yellow Balls movement on X axis, we need to add 2 different balls (Green and Red) in the game. Green Balls will be more fast then yellow one and their score will be 20 points. Red balls will be the fastest balls in the game and if catched will decrease players Life by 1.

In FlyingFishView java Class:
- Add int variables for Green and Red balls as for yellow one.
- Add and Set a new Paint (greenPaint and redPaint).
- Draw the ball/s on canvas.
- Add a method() hitBallChecker to catch the balls according to fish and balls position.